pyproject.toml
[project]
name = "villajets"
version = "0.1.0"
description = ""
authors = [
    {name = "Jose Garcia Escandón",email = "jose.g@covergenius.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-unfold (>=0.57.0,<0.58.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-admin-interface (>=0.30.0,<0.31.0)",
    "filetype (>=1.2.0,<2.0.0)",
    "dateparser (>=1.2.1,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pycountry (>=24.6.1,<25.0.0)",
    "pdfminer (>=20191125,<20191126)",
    "pdfminer-six (>=20250506,<20250507)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "passporteye (>=2.2.2,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-colorfield (>=0.14.0,<0.15.0)",
    "imageio (>=2.37.0,<3.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "packaging (>=25.0,<26.0)",
    "six (>=1.17.0,<2.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "tzlocal (>=5.3.1,<6.0.0)",
    "cryptography (>=45.0.4,<46.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "cffi (>=1.17.1,<2.0.0)",
    "pycparser (>=2.22,<3.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "threadpoolctl (>=3.6.0,<4.0.0)",
    "networkx (>=3.5,<4.0)",
    "fonttools (>=4.58.2,<5.0.0)",
    "kiwisolver (>=1.4.8,<2.0.0)",
    "cycler (>=0.12.1,<0.13.0)",
    "lazy-loader (>=0.4,<0.5)",
    "contourpy (>=1.3.2,<2.0.0)",
    "tifffile (>=2025.6.1,<2026.0.0)",
    "text-unidecode (>=1.3,<2.0)",
    "django-countries (>=7.6.1,<8.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
faker = "^37.3.0"

